requirements.txt
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0
brotlipy==0.7.0
cchardet==2.1.7
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
cmake==3.27.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
crcmod==1.7
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.0
dataclasses-json==0.6.1
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
faiss-gpu==1.7.2
fastapi==0.103.2
fastjsonschema==2.18.0
ffmpy==0.3.1
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
fuzzywuzzy==0.18.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
gradio==3.45.2
gradio_client==0.5.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.5
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
InstructorEmbedding==1.0.1
ipykernel==6.25.2
ipython==8.16.0
isoduration==20.11.0
jedi==0.19.0
Jinja2 @ file:///croot/jinja2_1666908132255/work
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langchain==0.0.305
langsmith==0.0.41
lark==1.1.7
lazy_loader==0.3
Levenshtein==0.22.0
lit==17.0.1
lmdb==1.4.1
loralib==0.1.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmengine==0.8.5
mmocr==1.0.1
model-index==0.1.11
monotonic==1.6
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx @ file:///croot/networkx_1690561992265/work
nltk==3.8.1
notebook_shim==0.2.3
numexpr==2.8.7
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp310-cp310-linux_x86_64.whl#sha256=fc2732718bc9e06a7b702492cb4f5afffe9671083930452d894377bf563464a3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.16.0
openai==0.28.1
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.26
ordered-set==4.1.0
orjson==3.9.7
oss2==2.17.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==3.10.0
posthog==3.0.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PyPika==0.48.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.22.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
rank-bm25==0.2.2
rapidfuzz==3.3.1
referencing==0.30.2
regex==2023.8.8
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.10.3
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.2
starlette==0.27.0
sympy @ file:///croot/sympy_1668202399572/work
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.14.0
tomli==2.0.1
toolz==0.12.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.3.3
tqdm==4.65.2
traitlets==5.10.1
transformers @ git+https://github.com/huggingface/transformers@0b192de1f353b0e04dad4813e02e2c672de077be
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1686163155763/work
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.7
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==11.0.3
xformers==0.0.22
xxhash==3.3.0
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
